api/pr_reviewer/api/requirements.txt
pandas
fastapi
uvicorn[standard]
pydantic
requests
loguru
pyyaml
matplotlib
wpipe
python-multipart